feat: support explicit formatter (#71)
## PR Checklist

- [x] Addresses an existing open issue: fixes #73
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/formatly/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/formatly/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

I'd like to be able to reuse `formatly` with an explicit formatter
instead of detecting. This options helps for cases where a top-level
tooling allows users to explicitly specify the formatter instead of
auto-detecting.

I think it'd be nice to also rename the accepted names to `"biome"`,
`"deno"`, `"dprint"`, and `"prettier"` to simplify the casing and
spacing, but I'd like to get thoughts on this first before making the
breaking change.